* development: (32 commits)
  fixed bug where bundler wasn't used on server to run schedueld jobs
  changing the end of a shift from UTC to localtime
  Removed unnecessary view
  Small changes for readability
  minor table renaming change
  fixed bug in profile repair task
  minor change to location report items view
  Made the change James suggested
  quick change, but the other branch was messy so I made this one, yay.
  Small change to rake wording
  manually adding Maria's fix for bug 381
  Adding fix to bug 420 manually
  Added a comment to the if statement
  Enabled department_config option functionality
  Changed periodically call remote frequency
  Preliminary code
  Function works
  Fixed index views
  user profile search works, made a few changes
  rake task by departments
  ...

+++ app/models/shift.rb
@@ -186,7 +186,7 @@ class Shift < ActiveRecord::Base
       return out.collect{|t| "The shift for "+t.to_message_name+" conflicts. Use wipe to fix."}.join(",")
     end
   end
-
+  
 
   #Used for activating calendars, check/wipe conflicts -Mike
   def self.check_for_conflicts(shifts, wipe)
@@ -259,6 +259,12 @@ class Shift < ActiveRecord::Base
       return number_report_items/shift_time
     end
   end
+  
+  #to enable the view of unscheduled shifts, a shift that lacks an end attribute is viewed as ending right now
+  # else, the end attribute is read (referenced in shifts_helper)
+  def end
+     read_attribute(:end).nil? ? Time.now : read_attribute(:end).localtime
+   end
 
   #a shift has been signed in to if it has a report
   # NOTE: this evaluates whether a shift is CURRENTLY signed in
@@ -297,21 +303,25 @@ class Shift < ActiveRecord::Base
   def combine_with_surrounding_shifts
     #if (shift_later = Shift.find(:first, :conditions => {:start => self.end, :user_id => self.user_id, :location_id => self.location_id, :calendar_id => self.calendar.id}))
     if (shift_later = Shift.find(:first, :include => :calendar, :conditions => ["start = ? AND user_id = ? AND location_id = ? AND calendars.active = ?", self.end, self.user_id, self.location_id, self.calendar.active?]))
-      self.end = shift_later.end
-      shift_later.sub_requests.each { |s| s.shift = self }
-      shift_later.destroy
-      self.save(false)
+      unless self.submitted?
+       self.end = shift_later.end
+       shift_later.sub_requests.each { |s| s.shift = self }
+       shift_later.destroy
+       self.save(false)
+      end
     end
     #if (shift_earlier = Shift.find(:first, :conditions => {:end => self.start, :user_id => self.user_id, :location_id => self.location_id, :calendar_id => self.calendar.id}))
     if (shift_earlier = Shift.find(:first, :include => :calendar, :conditions => ["end = ? AND user_id = ? AND location_id = ? AND calendars.active = ?", self.start, self.user_id, self.location_id, self.calendar.active?]))
-      self.start = shift_earlier.start
-      shift_earlier.sub_requests.each {|s| s.shift = self}
-      unless shift_earlier.report.nil?
-        shift_earlier.report.shift = nil
-        shift_earlier.report.save! #we have to disassociate the report first, or it will be destroyed too
-        self.report = shift_earlier.report
-        shift_earlier.report = nil
+      unless shift_earlier.submitted?
+        self.start = shift_earlier.start
+        shift_earlier.sub_requests.each {|s| s.shift = self}
+        unless shift_earlier.report.nil?
+          shift_earlier.report.shift = nil
+          shift_earlier.report.save! #we have to disassociate the report first, or it will be destroyed too
+          self.report = shift_earlier.report
+          shift_earlier.report = nil
       end
+    end
       self.signed_in = shift_earlier.signed_in
       shift_earlier.destroy
       self.save(false)
